+++ src/components/graph/Graph.tsx
@@ -8,9 +8,10 @@ import {
   XAxis,
   YAxis,
 } from 'recharts';
-import { format, parseISO } from 'date-fns/esm';
+import { differenceInDays, format, parseISO } from 'date-fns/esm';
 import { CurveType } from 'recharts/types/shape/Curve';
 import { AxisDomain } from 'recharts/types/util/types';
+import { getEvenlySpacedDates } from '../../util/Dates';
 
 export function getIdealStep(diffInDays: number, numUniqueYears: number) : number {
   if (diffInDays <= 7) {
@@ -82,8 +83,6 @@ export type GraphProps = {
   containerHeight: number;
   containerClassName?: string;
   charts: GraphChart[];
-  dateFormat: string;
-  ticks: string[];
   tickTextColor: string;
   linearGradients?: React.SVGProps<SVGLinearGradientElement>[];
   CustomTooltip?: JSX.Element;
@@ -103,8 +102,6 @@ export default function Graph(props: GraphProps) {
     containerHeight,
     containerClassName,
     charts,
-    dateFormat,
-    ticks,
     tickTextColor,
     linearGradients,
     CustomTooltip,
@@ -117,6 +114,17 @@ export default function Graph(props: GraphProps) {
     yAxisDomain,
     setIsActive,
   } = props;
+  
+  const dates = data.map((d: any) => d.x) as string[];
+  const updatedTo = new Date(dates[dates.length - 1]);
+  const updatedFrom = new Date(dates[0]);
+  const numberOfUniqueYears = new Set(
+    dates.map((d) => parseISO(d).getFullYear())
+  ).size;
+  const diffInDays = differenceInDays(updatedTo, updatedFrom);
+  const step = getIdealStep(diffInDays, numberOfUniqueYears);
+  const dateFormat = getIdealDateFormat(diffInDays);
+  const ticks = getEvenlySpacedDates(dates, step).slice(1);
 
   return (
     <CustomizedResponsiveContainer
